+++ packages/adhere-util-domain/lib/index.js
@@ -1,0 +1,2 @@
+import Domain from"./domain";export default Domain;
+//# sourceMappingURL=index.js.map