+++ packages/adhere/lib/domain/index.js
@@ -1,0 +1,2 @@
+import Model from"@baifendian/adhere-util-domain";export default Model;
+//# sourceMappingURL=index.js.map